third_party/blink/web_tests/platform/generic/touchadjustment/scroll-offset-expected.txt

Test touch adjustment when main frame is scroll offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS adjusted node was A#a0.
PASS adjusted node was A#a1.
PASS adjustedPoint is non-null.
PASS adjustedNode is e.a0
PASS adjustedPoint is non-null.
PASS adjustedNode is e.a1
PASS successfullyParsed is true

TEST COMPLETE
